+++ test/form/samples/export-globals/Number.js
@@ -1,1 +1,10 @@
-export { isNaN, isFinite, parseFloat, Number }
+const localIsNan = isNan;
+const localIsFinite = isFinite;
+const localParseFloat = parseFloat;
+const localNumber = Number;
+export {
+	localIsNan as isNaN,
+	localIsFinite as isFinite,
+	localParseFloat as parseFloat,
+	localNumber as Number
+};