+++ test/form/samples/export-globals/_expected/iife.js
@@ -1,9 +1,11 @@
 var myBundle = (function (exports) {
 	'use strict';
 
-	const isNaN$1 = isNaN;
+	const localIsNan = isNan;
 
-	exports.isNaN = isNaN$1;
+	const isNaN = localIsNan;
+
+	exports.isNaN = isNaN;
 
 	return exports;