+++ test/form/samples/export-globals/_expected/umd.js
@@ -4,9 +4,11 @@
 	(global = global || self, factory(global.myBundle = {}));
 }(this, function (exports) { 'use strict';
 
-	const isNaN$1 = isNaN;
+	const localIsNan = isNan;
 
-	exports.isNaN = isNaN$1;
+	const isNaN = localIsNan;
+
+	exports.isNaN = isNaN;
 
 	Object.defineProperty(exports, '__esModule', { value: true });